+++ src/main/java/br/com/meli/PIFrescos/controller/dtos/InboundOrderUpdateDTO.java
@@ -1,6 +1,5 @@
-package br.com.meli.PIFrescos.dtos;
+package br.com.meli.PIFrescos.controller.dtos;
 
-import br.com.meli.PIFrescos.controller.dtos.BatchDTO;
 import br.com.meli.PIFrescos.models.InboundOrder;
 import br.com.meli.PIFrescos.models.Section;
 import lombok.AllArgsConstructor;
